requirements.txt
opencv-python==4.8.1.78
pandas==2.1.4
numpy==1.26.2
streamlit==1.29.0
matplotlib==3.8.2
scikit-learn==1.3.2
Pillow==10.1.0 